Move MySQL outbox implementation into `outbox` subpackage.
