+++ src/axmysql/outbox/pkg.go
@@ -1,0 +1,3 @@
+// Package outbox provides MySQL-specific implementations of the
+// interfaces in Ax's top-level "outbox" package.
+package outbox
